# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BoardRecord {
  id: number;
  code: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  stock_length?: number | null;
  /** @nullable */
  stock_width?: number | null;
}

export interface BoardInput {
  code: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  stock_length: number | null;
  /** @nullable */
  stock_width: number | null;
}

export interface BoardUpdate {
  code?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  stock_length?: number | null;
  /** @nullable */
  stock_width?: number | null;
}

export interface NotFound {
  error: string;
}

export type GetBoardByCodeParams = {
/**
 * Optional board code to look up (e.g. "3mm MDF" or "FBCT540X60"). Omit to list all boards.
 */
code?: string;
};

export type DeleteBoard200 = {
  ok: boolean;
};