# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: boards
    description: Board/material records
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /boards:
    get:
      operationId: getBoardByCode
      tags: [boards]
      summary: Look up a board by code
      description: Returns stock dimensions for a board code if found
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
          description: Optional board code to look up (e.g. "3mm MDF" or "FBCT540X60"). Omit to list all boards.
      responses:
        "200":
          description: Board found
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/BoardRecord"
                  - type: array
                    items:
                      $ref: "#/components/schemas/BoardRecord"
        "404":
          description: Board not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFound"
    post:
      operationId: createBoard
      tags: [boards]
      summary: Create a board stock record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BoardInput"
      responses:
        "201":
          description: Board created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BoardRecord"
        "400":
          description: Invalid board
        "409":
          description: Board code already exists
  /boards/{id}:
    put:
      operationId: updateBoardById
      tags: [boards]
      summary: Update a board stock record by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BoardUpdate"
      responses:
        "200":
          description: Board updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BoardRecord"
        "404":
          description: Board not found
        "409":
          description: Board code already exists
    delete:
      operationId: deleteBoard
      tags: [boards]
      summary: Delete a board stock record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Board deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required: [ok]
        "404":
          description: Board not found
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    BoardRecord:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        description:
          type: ["string", "null"]
        stock_length:
          type: ["integer", "null"]
        stock_width:
          type: ["integer", "null"]
      required:
        - id
        - code
    BoardInput:
      type: object
      properties:
        code:
          type: string
        description:
          type: ["string", "null"]
        stock_length:
          type: ["integer", "null"]
        stock_width:
          type: ["integer", "null"]
      required:
        - code
        - stock_length
        - stock_width
    BoardUpdate:
      type: object
      minProperties: 1
      properties:
        code:
          type: string
        description:
          type: ["string", "null"]
        stock_length:
          type: ["integer", "null"]
        stock_width:
          type: ["integer", "null"]
    NotFound:
      type: object
      properties:
        error:
          type: string
      required:
        - error