# lib/api-zod/src/generated/types/boardUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BoardUpdate {
  code?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  stock_length?: number | null;
  /** @nullable */
  stock_width?: number | null;
}